LangChain/requirements.txt
langchain
python-dotenv
langchain-community
pypdf
bs4
arxiv
PyMuPDF
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain-ollama
numpy
scipy
sentence_transformers
langchain_huggingface
tqdm
faiss-cpu
langchain_chroma
beautifulsoup4
streamlit
